Разнес DTO и DAO по разным пакетам
