+++ organizer/src/main/java/inno/dao/TestDAOImpl.java
@@ -1,71 +1,46 @@
 package inno.dao;
 
 import classes.*;
+import connectionmanager.ConnectionManager;
+import connectionmanager.ConnectionManagerPostgresImpl;
 import exceptions.TestDAOException;
+import inno.dto.TestDTO;
+import inno.exceptions.TestDTOException;
 import interfaces.dao.TestDAO;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
 
-public class TestDAOImpl implements TestDAO {
-    @Override
-    public List<Test> getTestsByTeacher(Teacher teacher) {
+public class TestDAOImpl {
+
+    private static ConnectionManager manager = ConnectionManagerPostgresImpl.getInstance();
+
+    public static TestDTO getById(int id) throws TestDTOException {
+        String sql = "SELECT t.id, t.status, tt.topic, tt.description, sc.id AS school_class_id " +
+                "FROM tests AS t " +
+                "JOIN test_templates AS tt ON t.test_template_id = tt.id " +
+                "JOIN school_classes AS sc ON t.school_class_id = sc.id ";
+
+        try {
+            Statement statement = manager.getConnection().createStatement();
+            ResultSet rs = statement.executeQuery(sql);
+            if (rs.next()) {
+                return new TestDTO(
+                        rs.getInt("id"),
+                        rs.getString("status"),
+                        rs.getString("topic"),
+                        rs.getString("description"),
+                        rs.getInt("school_class_id")
+                );
+            }
+        } catch (SQLException e) {
+            throw new TestDTOException(e);
+        }
         return null;
     }
 
-    @Override
-    public List<Test> getTestsByTeacherDuringPeriod(Teacher teacher, LocalDate begin, LocalDate end) {
-        return null;
-    }
-
-    @Override
-    public List<Work> getWorks() {
-        return null;
-    }
-
-    @Override
-    public List<Question> getQuestions() {
-        return null;
-    }
-
-    @Override
-    public boolean addQuestions(Collection<Question> questions) {
-        return false;
-    }
-
-    @Override
-    public boolean assignToClass(SchoolClass schoolClass) {
-        return false;
-    }
-
-    @Override
-    public boolean assignDate(LocalDate date) {
-        return false;
-    }
-
-    @Override
-    public boolean addWork(Work test) {
-        return false;
-    }
-
-    @Override
-    public boolean markAbsent(Student student) {
-        return false;
-    }
-
-    @Override
-    public boolean isFullyLoaded() {
-        return false;
-    }
-
-    @Override
-    public boolean saveAsTemplate(Test event) {
-        return false;
-    }
-
-    @Override
-    public Test getById(int id) throws TestDAOException {
-        return null;
-    }
 }
