+++ organizer/src/main/java/inno/servlets/OrganizerServlet.java
@@ -1,5 +1,6 @@
 package inno.servlets;
 
+import inno.dao.TestDAOImpl;
 import inno.dto.TestDTO;
 import inno.exceptions.TestDTOException;
 
@@ -13,12 +14,12 @@ public class OrganizerServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("cp1251");
-        int test_id = req.getParameter("id") != null ?
-                Integer.parseInt(req.getParameter("id")) : 0;
+        int test_id = req.getParameter("test_id") != null ?
+                Integer.parseInt(req.getParameter("test_id")) : 0;
 
         if (test_id > 0) {
             try {
-                req.setAttribute("testDTO", TestDTO.getById(test_id));
+                req.setAttribute("testDTO", TestDAOImpl.getById(test_id));
             } catch (TestDTOException e) {
                 throw new ServletException(e);
             }
